requirements.txt
fastapi[standard]==0.115.7
ruff==0.9.3
pytest==8.3.4
pytest-cov==6.0.0
taskipy==1.14.1
sqlalchemy==2.0.37
pydantic-settings==2.7.1
alembic==1.14.1
pwdlib[argon2]
python-multipart
pyjwt==2.10.1
pytz